Moved custom tasks to handwire.c as _kb()
